feat(frontend): creates route system and improves architecture
creates a routing system with some provisional routes, furthermore improves the organization of
directories, creates a 'types' folder to store typing files; the 'pages' folder for storing complete
pages ('404 - not found' page was created); the 'routes' folder to handle the routing. finally, it
fixes and simplifies the api request functions.